+++ R/htmlTable_theme.R
@@ -198,12 +198,17 @@ prGetThemeListObject <- function(theme_name = c("standard", "Google docs", "blan
       ),
 
       css.cell = getOption("htmlTable.css.cell", default = "margin: 0; padding: 0;"),
-      css.cgroup = getOption("htmlTable.css.cgroup", default = "margin: 0; padding: 0;"),
-      css.header = getOption("htmlTable.css.header", default = "margin: 0; padding: 0; font-weight: 900"),
+      css.cgroup = getOption("htmlTable.css.cgroup", default = "margin: 0; padding: 0; vertical-align: middle;"),
+      css.header = getOption("htmlTable.css.header", default = "margin: 0; padding: 0; font-weight: 900; vertical-align: middle;"),
       css.header.border_bottom = getOption("htmlTable.css.header.border_bottom", default = "border-bottom: 1px solid grey"),
 
       css.class = getOption("htmlTable.css.class", default = "gmisc_table"),
       css.table = getOption("htmlTable.css.table", default = "margin-top: 1em; margin-bottom: 1em;"),
+
+      spacer.celltype = "double_cell",
+      spacer.css.cgroup.bottom.border = "1px solid white",
+      spacer.content = "",
+      spacer.css = "width: 2px;",
 
       # Positions
       pos.rowlabel = "bottom",