+++ NEWS.md
@@ -2,7 +2,8 @@ NEWS for the htmlTable package
 
 Changes fro 2.1.0
 -----------------
-* Added options for how the empty spacer cell appears (see addHtmlTableStyle spacer.* options)
+* Added options for how the empty spacer cell appears (see `addHtmlTableStyle()` spacer.* options)
+* Fixed ignored `align.header`
 
 Changes for 2.0.1
 -----------------